chore: combine atan2.default and atan2.out test files
